- moved some drawing code out of OpenGLFrameBuffer class.
git-svn-id: http://mancubus.net/svn/hosted/gzdoom/trunk@371 b0f79afe-0144-0410-b225-9a4edf0717df

+++ src/gl/common/glc_bsp.cpp
@@ -222,7 +222,7 @@ static inline void RenderThings(subsector_t * sub, sector_t * sector)
 
 static void DoSubsector(subsector_t * sub)
 {
-	int i;
+	unsigned int i;
 	sector_t * sector;
 	sector_t * fakesector;
 	sector_t fake;
@@ -247,7 +247,7 @@ static void DoSubsector(subsector_t * sub)
 
 	// [RH] Add particles
 	//int shade = LIGHT2SHADE((floorlightlevel + ceilinglightlevel)/2 + r_actualextralight);
-	for (i = ParticlesInSubsec[sub-subsectors]; i != NO_PARTICLE; i = Particles[i].snext)
+	for (i = ParticlesInSubsec[DWORD(sub-subsectors)]; i != NO_PARTICLE; i = Particles[i].snext)
 	{
 		GLRenderer->ProcessParticle(&Particles[i], fakesector);
 	}
